Move implementations to independent packages
This commit moves the text-based format contributed by Benjy from
Twitter to the `text` package and the new protobuf binary format to the
`binary` package.

This helps keep everything cleaner.
